01_Primeros_pasos_Python/06_Curso_Django/practica/cofee_shop/requirements.txt
Django==5.2
pillow==11.2.1
crispy-tailwind==1.0.3
djangorestframework==3.16.0
django-environ==0.12.0
psycopg2-binary==2.9.10